requirements.txt

pyodbc
pandas
numpy
apache-airflow-providers-odbc 
apache-airflow-providers-microsoft-mssql 
sqlalchemy